+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Go to definition finds annotated Python assignments (`NAME: Final[int] = ...`).
+
 ## [0.1.1] - 2026-09-12
 
 ### Fixed